move all fixtures into their respective domain folders
