+++ app/domains/fluggastrechte/services/pdf/sections/reason/factsOfCases/detailedReason/__test__/addMultiplePersonsText.test.ts
@@ -1,4 +1,4 @@
-import { userDataMock } from "tests/factories/fluggastrechte/userDataMock";
+import { userDataMock } from "~/domains/fluggastrechte/services/pdf/__test__/userDataMock";
 import {
   mockPdfKitDocument,
   mockPdfKitDocumentStructure,
